Remove "Candidate reviewing application" spec
At a price of not checking that the "Complete section" links work, move
the assertions on the presence and absence of banners into "Candidate
submitting application" spec and knock ~15 seconds off the test run.

Replace that spec with candidate_reviewing_an_incomplete_application,
which focusses on the empty review state and checks only that the links
are present - we trust the components otherwise.
